fix:コンポーネント構成見直し
